New Codebase
This shit is amazing. Seriously, I can't even tell you  how much better organized this code is. It's even commented (mostly). It's like a work of art. It's like a Picasso painting, in that working on it made me want to chop off part of my own body. But it's done, and it's awesome.